% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_report_rtf.R
\name{rtf_addreporttitle}
\alias{rtf_addreporttitle}
\title{在RTF文档中插入与“rtf_start”function中相同的reporttitle文本的function}
\usage{
rtf_addreporttitle()
}
\description{
在RTF文档中插入与“rtf_start”function中相同的reporttitle文本的function，注意调用时需要最后面加上“()”，如：rtf_addreporttitle () 。
}
\examples{
library(ggplot2)
mt <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point()
rtf_start(reportname = "null_rtf.rtf",reporttitle = c("XXXX","Report for IDMC","Sponsor",""))
rtf_addtextline(text="Table 1：RTF file output table test",style="title")
rtf_pagebreak()
rtf_addreporttitle ()
rtf_plot(mt,width = 10, height = 5.4)
rtf_end()

}
