% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_report_rtf.R
\name{rtf_addtable}
\alias{rtf_addtable}
\title{在RTF文档中插入表格的function}
\usage{
rtf_addtable(
  indat,
  varlist = NULL,
  norecord = "No Observed Value Was Reported.",
  protectspecialchars = FALSE,
  perpage_obs = NA
)
}
\arguments{
\item{indat}{指定在RTF文档中添加的表格，必须是data.frame}

\item{varlist}{包含信息：var/header/width/merge header/…/\\align\\order，具体见使用说明文档。
需列出的变量及其显示名称，但是需要注意的是，如果需要指定“\\align\\order”部分必须在其前面加上“/”,例如“/\\align\\order”.}

\item{norecord}{指定当导入的数据集观测数为0的时候输出的文本.}

\item{protectspecialchars}{指定是否对特殊文本进行保护。设置为TRUE时，文本中的RTF码将视为普通文本.}

\item{perpage_obs}{指定每页输出行数，当设置为NA时表示直接把数据集观测数作为每页行数，当指定具体行数num时按指定的具体行数num分页，当指定分页变量（pagebrak varibale）时则按分页变量值分页（注意变量名需要加上引号）.}
}
\description{
在RTF文档中插入表格的function，具体见使用说明文档。
}
\examples{
library(ggplot2)
mt <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point()
rtf_start(reportname = "null_rtf.rtf",reporttitle = c("XXXX","Report for IDMC","Sponsor",""))
rtf_addtextline(text="Table 1：RTF file output table test",style="title")
rtf_addtable(adae1,varlist=c("SUBJID/label 01/10/\\\\l\\\\order"
                            ,"TRT01P/label 02/10"
                            ,"AETERM/label 03/10/header\%\\\\test02/header zz/\\\\c"
                            ,"AESOC/label 04/10/header\%\\\\test02/header zz/\\\\r"))
rtf_end()

}
