% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_report_rtf.R
\name{rtf_addtextline}
\alias{rtf_addtextline}
\title{在RTF文档中插入文本行的function}
\usage{
rtf_addtextline(
  text = NULL,
  fontsize = 10.5,
  style = "normal",
  align = "left",
  bold = FALSE,
  protectspecialchars = FALSE,
  inheader = NULL
)
}
\arguments{
\item{text}{指定添加进RTF的文本。}

\item{fontsize}{指定字体大小.}

\item{style}{指定输出的样式，title样式会生成导航栏书签，footnote时将文本插入页脚。}

\item{align}{指定文本对齐方式.}

\item{bold}{指定是否加粗文本.}

\item{protectspecialchars}{指定是否对特殊文本进行保护。设置为TRUE时，文本中的RTF码将视为普通文本.}

\item{inheader}{指定是否将文本插入header(页眉)中，仅对style=title时有效.}
}
\description{
在RTF文档中插入文本行的function。
}
\examples{
library(ggplot2)
mt <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point()
rtf_start(reportname = "null_rtf.rtf",reporttitle = c("XXXX","Report for IDMC","Sponsor",""))
rtf_addtextline(text="Table 1：RTF file output table test",style="title")
rtf_pagebreak()
rtf_plot(mt,width = 10, height = 5.4)
rtf_end()

}
