% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_report_rtf.R
\name{rtf_end}
\alias{rtf_end}
\title{输出RTF文档的结束位置function}
\usage{
rtf_end()
}
\description{
输出RTF文档的结束位置function，注意调用时需要最后面加上“()”，如：rtf_end()。
}
\examples{
library(ggplot2)
mt <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point()
rtf_start(reportname = "null_rtf.rtf",reporttitle = c("XXXX","Report for IDMC","Sponsor",""))
rtf_addtextline(text="Table 1：RTF file output table test",style="title")
rtf_pagebreak()
rtf_plot(mt,width = 10, height = 5.4)
rtf_end()

}
