% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_report_rtf.R
\name{rtf_plot}
\alias{rtf_plot}
\title{在RTF文档中插入图片的function}
\usage{
rtf_plot(
  x,
  imagefmt = "emf",
  width = 7,
  height = 7,
  bg = "transparent",
  fg = "black",
  pointsize = 12,
  family = "Arial",
  coordDPI = 1000,
  custom.lty = TRUE,
  emfPlus = TRUE,
  emfPlusFont = FALSE,
  emfPlusRaster = FALSE,
  emfPlusFontToPath = FALSE,
  res = 300,
  symbolfamily = "default"
)
}
\arguments{
\item{x}{一个绘图对象}

\item{imagefmt}{指定输出图片的格式，可选emf/png/bmp}
}
\description{
在RTF文档中插入图片的function，参数除了x是绘图对象object，其他都是devEMF包中emf function或者grDevices包（png/bmp）function参数，所以具体参数参考devEMF包(emf)或grDevices包（png/bmp）说明。
}
\examples{
library(ggplot2)
mt <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point()
rtf_start(reportname = "null_rtf.rtf",reporttitle = c("XXXX","Report for IDMC","Sponsor",""))
rtf_addtextline(text="Table 1：RTF file output table test",style="title")
rtf_pagebreak()
rtf_plot(mt,width = 10, height = 5.4)
rtf_end()

}
