% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_report_rtf.R
\name{rtf_start}
\alias{rtf_start}
\title{输出RTF文档的开始位置function}
\usage{
rtf_start(
  outpath = ".",
  reportname = "report_rtf.rtf",
  reporttitle = c("Prject name", "Report for", "Sponsor", ""),
  papersize = "A4",
  fontsize = 10.5,
  font = "SimSun/Times New Roman",
  protectspecialchars = FALSE,
  marglrtb = c(1, 1, 1, 1),
  orientation = "landscape",
  head_footery = c(0.5, 0.5),
  bodytitle = TRUE
)
}
\arguments{
\item{outpath}{RTF文件输出路径}

\item{reportname}{指定输出的RTF名称（必须带上扩展名“.rtf”）}

\item{reporttitle}{指定页眉信息（包含项目信息）}

\item{papersize}{指定纸张大小，只支持A4或LETTER.}

\item{fontsize}{指定字体大小.}

\item{font}{指定输出字体，支持双字体（字体间用“/”隔开）}

\item{protectspecialchars}{指定是否对特殊文本进行保护。设置为TRUE时，文本中的RTF码将视为普通文本.}

\item{marglrtb}{指定页边距大小，单位为in，页边距设置顺序依次为左右上下.}

\item{orientation}{指定纸张方向，PORTRAIT=纵向，landscape=横向.}

\item{head_footery}{指定页眉文本与页边距离}

\item{bodytitle}{指定页眉文本放在正文还是header中}
}
\description{
输出RTF文档的开始位置function，可以设置一些常见文档页面参数，如纸张大小、页面方向、页边距等（具体见下方参数）。
}
\examples{
library(ggplot2)
mt <- ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point()
rtf_start(reportname = "null_rtf.rtf",reporttitle = c("XXXX","Report for IDMC","Sponsor",""))
rtf_addtextline(text="Table 1：RTF file output table test",style="title")
rtf_pagebreak()
rtf_plot(mt,width = 10, height = 5.4)
rtf_end()

}
